% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco1.R
\name{.make.bisseco.culdesac.tfgy}
\alias{.make.bisseco.culdesac.tfgy}
\title{Generates an epidemiological history as input for the tree simulation functions}
\usage{
.make.bisseco.culdesac.tfgy(
  mu,
  gamma,
  alpha,
  omega,
  yscale,
  mh,
  maxst,
  Net,
  pa,
  res = 200
)
}
\description{
See _phydynR_ package for details on output format
}
