% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco1.R
\name{fitbisseco}
\alias{fitbisseco}
\title{Fits the parameters of a BiSSeCo model by maximum likelihood}
\usage{
fitbisseco(
  tr,
  isvariant,
  Tg,
  mu,
  Net = NULL,
  theta0 = log(c(alpha = 15, omega = 0.95, yscale = 1)),
  augment_likelihood = TRUE,
  optim_parms = list(),
  mlesky_parms = list(),
  res = 200,
  ...
)
}
\arguments{
\item{tr}{an ape::phylo representing a time-scaled phylogeny. These can be computed with the `treedater` package}

\item{isvariant}{vector of type boolean with length equal to ape::Ntip(tr). Each element is TRUE if the corresponding element in tr$tip.label is a variant type}

\item{Tg}{Generation time, i.e. the mean time elapsed between generations. Units of this variable should match those used in tr$edge.length and 1/mu, e.g. days or years}

\item{mu}{Molecular clock rate of evolution}

\item{Net}{NULL or a matrix with two columns giving the effective population size. If NULL, the effective population size will be computed with the `mlesky` package. The first column should be time since some point in the past, and the second column should be an estimate of the effective population size. Time zero should correspond to the root of the tree}

\item{theta0}{Initial guess of (log) parameter values: mu, omega, and Ne scale.}

\item{augment_likelihood}{if TRUE (default), will combine the coalescent likelihood with a binomial likelihood of sampling variants or ancestral types under the assumption of random sampling and mutation-selection balance}

\item{optim_parms}{optional list of parameters passed to optim when fitting the model}

\item{mlesky_parms}{optional list of parameters passed to mlesky::mlskygrid if estimating Ne(t)}

\item{res}{Integer number time steps used in coalescent likelihood}

\item{...}{Additional parameters are passed to phydynR::colik}
}
\description{
This method estimates a pair of selection coefficients from pathogen phylogenies representing differences in transmissibility and differences in within-host fitness. 
Pathogen phylogenies are assumed to be reconstructed from population-based random samples of pathogen genomes and at most one sequence per host. 
Phylogenies should be time-scaled, and an estimate of the molecular clock rate of evolution should be provided, such as estimated with the `treedater` package.
}
