% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco-package.R
\docType{package}
\name{musseco-package}
\alias{musseco}
\alias{musseco-package}
\title{musseco: Multi-State Speciation and Extinction Coalescent}
\description{
Methods for simulating genealogies and estimating parameters for a speciation and extinction coalescent process such that 1) disadvantageous alleles appear continuously in a population and are present in mutation/selection balance; 2) non-equilibrium dynamics of advantageous alleles which appear continuously and grow in frequency.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://emvolz-phylodynamics.github.io/musseco/}
  \item \url{https://github.com/emvolz-phylodynamics/musseco}
  \item Report bugs at \url{https://github.com/emvolz-phylodynamics/musseco/issues}
}

}
\author{
\strong{Maintainer}: Erik Volz \email{erik@erikvolz.info} (\href{https://orcid.org/0000-0001-6268-8937}{ORCID})

Other contributors:
\itemize{
  \item Kieran O. Drake \email{k.o.drake@gmail.com} (\href{https://orcid.org/0000-0002-1280-9523}{ORCID}) [contributor]
  \item Fabricia F. Nascimento \email{thednainus@yahoo.com} (\href{https://orcid.org/0000-0001-9426-6835}{ORCID}) [contributor]
}

}
\keyword{internal}
