% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco.R, R/musseco.R.tmp.R
\name{pwt_mutsel_balance}
\alias{pwt_mutsel_balance}
\title{Code for simulating genealogies for a "Binary state speciation and extinction model" (BiSSE)
Trees are simulated using a structured coalescent framework implemented in the phydynR package
In this model, a wild type mutates at a constant rate to the mutant type which has lower transmission fitness.}
\usage{
pwt_mutsel_balance(mu, omega)

pwt_mutsel_balance(mu, omega)
}
\arguments{
\item{mu}{Rate of de novo mutation wt -> mutant}

\item{omega}{Relative transmission fitness (<1) of the mutant type}
}
\value{
Scalar frequency of the wild type

Scalar frequency of the wild type
}
\description{
Code for simulating genealogies for a "Binary state speciation and extinction model" (BiSSE)
Trees are simulated using a structured coalescent framework implemented in the phydynR package
In this model, a wild type mutates at a constant rate to the mutant type which has lower transmission fitness.

Code for simulating genealogies for a "Binary state speciation and extinction model" (BiSSE)
Trees are simulated using a structured coalescent framework implemented in the phydynR package
In this model, a wild type mutates at a constant rate to the mutant type which has lower transmission fitness.
}
\author{
Erik M Volz <erik.volz@gmail.com>
}
