% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco.R
\name{simulate_bisseco}
\alias{simulate_bisseco}
\title{Simulate a structured coalescent tree for the 'Binary State Speciation and Extinction Model'}
\usage{
simulate_bisseco(
  mu,
  omega,
  maxHeight,
  Net,
  sampleTimes,
  sampleStates,
  res = 200,
  ...
)
}
\arguments{
\item{mu}{Rate of de novo mutation ancestral -> variant}

\item{omega}{Relative transmission fitness (<1) of the variant type}

\item{maxHeight}{Maximum time in the past to simulate}

\item{Net}{Effective population size through time. This is stored as a two column
matrix, such that the first column is time and the second column is population size}

\item{sampleTimes}{A vector of sample times}

\item{sampleStates}{A matrix describing the state (ancestral or variant) for each sample}

\item{res}{Number of time points to simulate}
}
\value{
A simulated genealogy in ape::phylo format
}
\description{
Simulate a structured coalescent tree for the 'Binary State Speciation and Extinction Model'
}
\examples{
set.seed(23) 
# Sample times: these will be distributed over 150 years to show role of heterochronous sampling
sts <- setNames( sample( seq( 300, 500, length=20 )), letters[1:20] )

# We will have the effective sample size increase linearly through time so that external branch lengths are longer: 
net <- cbind( seq(0, 501, length=500),  seq(1,200,length=500)  )

# Sample states: arranged in a matrix where each row gives the probability of being wildtype or variant
ssts <- cbind( ancestral = c( rep(1,15), rep(0, 5))
 , variant = c( rep( 0, 15), rep(1, 5 ) ) 
 )
rownames( ssts ) <- letters[1:20]

# Scenario A: Low denovo mutation rate and relatively high transmission fitness
tr1 = simulate_bisseco( .001, .995
 , maxHeight = 500
 , Net = net
 , res = 500
 , sampleTimes = sts
 , sampleStates = ssts 
) 

# Scenario B: High de novo mutation rate and relatively low transission fitness
tr2 = simulate_bisseco( .05, .75
 , maxHeight = 500
 , Net = net 
 , res = 500
 , sampleTimes = sts
 , sampleStates = ssts 
) 

# Plot the annotated trees for both scenarios
p1 <-bisseco_plot( tr1, legend=TRUE) 
p2 <-bisseco_plot( tr2, legend=TRUE)
}
