% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco.R
\name{bisseco_plot}
\alias{bisseco_plot}
\title{Make a plot of a bisseco tree that shows ancestral states at internal nodes estimated using a simple discrete trait analysis}
\usage{
bisseco_plot(tr, legend = FALSE)
}
\arguments{
\item{tr}{A tree (phydynR::DatedTree) simulated using simulate_bisseco}
}
\description{
Make a plot of a bisseco tree that shows ancestral states at internal nodes estimated using a simple discrete trait analysis
}
