% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco.R
\name{.make.bisseco.culdesac.tfgy}
\alias{.make.bisseco.culdesac.tfgy}
\title{Generates an epidemiological history as input for the tree simulation functions}
\usage{
.make.bisseco.culdesac.tfgy(mu, omega, yscale, mh, maxst, Net, res = 200)
}
\arguments{
\item{mu}{Rate of de novo mutation ancestral -> variant}

\item{omega}{Relative transmission fitness (<1) of the variant type}

\item{mh}{Maximum time in the past to simulate}

\item{maxst, }{maximum sample time ; time of last tip in the tree}

\item{Net}{Effective population size through time. This is stored as a two column
matrix, such that the first column is time (forward, since some point in the past) and the second column is population size}

\item{res}{Number of time points to simulate}
}
\description{
See _phydynR_ package for details on output format
}
