% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco.R
\name{fitbisseco}
\alias{fitbisseco}
\title{Fits the parameters of a BiSSeCo model by maximum likelihood}
\usage{
fitbisseco(
  tr,
  isvariant,
  Net = NULL,
  theta0 = log(c(0.05, 0.95, 1)),
  augment_likelihood = TRUE,
  optim_parms = list(),
  mlesky_parms = list()
)
}
\arguments{
\item{tr}{an ape::phylo representing a time-scaled phylogeny. These can be computed with the `treedater` package}

\item{isvariant}{vector of type boolean with length equal to ape::Ntip(tr). Each element is TRUE if the corresponding element in tr$tip.label is a variant type}

\item{Net}{NULL or a matrix with two columns giving the effective population size. If NULL, the effective population size will be computed with the `mlesky` package. The first column should be time since some point in the past, and the second column should be an estimate of the effective population size. Time zero should correspond to the root of the tree}

\item{theta0}{Initial guess of (log) parameter values: mu, omega, and Ne scale.}

\item{augment_likelihood}{if TRUE (default), will combine the coalescent likelihood with a binomial likelihood of sampling variants or ancestral types under the assumption of random sampling and mutation-selection balance}

\item{optim_parms}{optional list of parameters passed to optim when fitting the model}

\item{mlesky_parms}{optional list of parameters passed to mlesky::mlskygrid if estimating Ne(t)}
}
\description{
This method will estimate a de novo mutation rate (mu) reflecting within-host fitness of a variant and between host relative fitness (omega) reflecting differences in transmissibility.
}
