% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musseco.R.tmp.R
\name{loglikelihood_bisseco}
\alias{loglikelihood_bisseco}
\title{Likelihood of the BiSSeCo model given rates of variant mutation, selection, a dated tree, and effective population size over time}
\usage{
loglikelihood_bisseco(mu_omega, dtr, Net, res = 200, ...)
}
\arguments{
\item{mu_omega}{a 2-vector containing parameters mu (mutation rate) and omega (selection coef)}

\item{dtr}{A phydynR::DatedTree}

\item{Net}{Effective population size through time. This is stored as a two column}

\item{maxHeight}{Maximum time in the past to simulate}
}
\value{
scalar log likelihood
}
\description{
Likelihood of the BiSSeCo model given rates of variant mutation, selection, a dated tree, and effective population size over time
}
\examples{
matrix, such that the first column is time and the second column is population size
set.seed( 23 ) 

# Sample times: these will be distributed over 150 years to show role of heterochronous sampling
sts <- setNames( sample( seq( 300, 500, length=20 )), letters[1:20] )

# We will have the effective sample size increase linearly through time so that external branch lengths are longer: 
net <- cbind( seq(0, 501, length=500),  seq(1,200,length=500)  )

# Sample states: arranged in a matrix where each row gives the probability of being wildtype or mutant
ssts <- cbind( wt = c( rep(1,15), rep(0, 5))
 , mutant = c( rep( 0, 15), rep(1, 5 ) ) 
 )
rownames( ssts ) <- letters[1:20]

# Scenario A: Low denovo mutation rate and relatively high transmission fitness
tr1 = simulate_bisseco( .001, .995
 , maxHeight = 500
 , Net = net
 , res = 500
 , sampleTimes = sts
 , sampleStates = ssts 
) 
print(  (l1 <- likelihood_bisseco( c(.005,.995), tr1, net ))  )
}
